namespace Core.Inputting.InputTree has been renamed to Core.Inputting.Tree
