+++ BeaverSoft.Texo.Core/Inputting/Tree/IInputTreeBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using BeaverSoft.Texo.Core.Configuration;
 
-namespace BeaverSoft.Texo.Core.Inputting.InputTree
+namespace BeaverSoft.Texo.Core.Inputting.Tree
 {
     public interface IInputTreeBuilder
     {
