+++ BeaverSoft.Texo.Core/Inputting/Tree/IInputTreeEvaluationStrategy.cs
@@ -1,4 +1,4 @@
-namespace BeaverSoft.Texo.Core.Inputting.InputTree
+namespace BeaverSoft.Texo.Core.Inputting.Tree
 {
     public interface IInputTreeEvaluationStrategy
     {
