+++ BeaverSoft.Texo.Core/Inputting/Tree/ParameterNode.cs
@@ -1,6 +1,6 @@
 ﻿using BeaverSoft.Texo.Core.Configuration;
 
-namespace BeaverSoft.Texo.Core.Inputting.InputTree
+namespace BeaverSoft.Texo.Core.Inputting.Tree
 {
     public class ParameterNode : BaseNode
     {
