+++ BeaverSoft.Texo.Test.Core/InputTreeEvaluationTests.cs
@@ -1,7 +1,7 @@
-﻿using BeaverSoft.Texo.Core.Configuration;
+using BeaverSoft.Texo.Core.Configuration;
 using BeaverSoft.Texo.Core.Environment;
 using BeaverSoft.Texo.Core.Inputting;
-using BeaverSoft.Texo.Core.Inputting.InputTree;
+using BeaverSoft.Texo.Core.Inputting.Tree;
 using StrongBeaver.Core.Services.Logging;
 using Xunit;
